refactor: move GameCard and related files to presentation layer
